Separate out core download configureDownloadButtons method
