+++ src/core/AdmissionCountsByType.js
@@ -30,7 +30,7 @@ import reduce from "lodash/fp/reduce";
 import toInteger from "lodash/fp/toInteger";
 
 import { COLORS, COLORS_FIVE_VALUES } from "../assets/scripts/constants/colors";
-import { configureDownloadButtons } from "../utils/downloads/downloads";
+import { configureDownloadButtons } from "./utils/configureDownloadButtons";
 import {
   filterDatasetByDistrict,
   filterDatasetBySupervisionType,
