+++ src/core/bars/utils.js
@@ -24,7 +24,7 @@ import toInteger from "lodash/fp/toInteger";
 import values from "lodash/fp/values";
 import upperCase from "lodash/fp/upperCase";
 
-import { configureDownloadButtons } from "../../utils/downloads/downloads";
+import { configureDownloadButtons } from "../utils/configureDownloadButtons";
 
 /**
  * Casts arguments to integer and sum it.
