+++ src/core/programming/FtrReferralsByRace.js
@@ -29,7 +29,7 @@ import {
   COLORS_FIVE_VALUES,
   COLORS,
 } from "../../assets/scripts/constants/colors";
-import { configureDownloadButtons } from "../../utils/downloads/downloads";
+import { configureDownloadButtons } from "../utils/configureDownloadButtons";
 import {
   filterDatasetBySupervisionType,
   filterDatasetByDistrict,
